Extract MatchApplier collaborator
Moves every write to app.transaction_categories, app.user_merchants,
app.user_categories, app.category_overrides, and app.categorization_rules
into a new categorization/applier.py, with WriteOutcome and
RuleCreationResult co-located. MatchApplier holds the Database handle plus
an injected AuditService so the audit-emitting paths (set_category,
set_category_in_active_txn, clear_category) stay together with
write_categorization — the bypass warning in set_category_in_active_txn
depends on living next to the guard it bypasses.

For the three write methods that historically composed in-place with
categorize_pending (create_merchant, create_rules, deactivate_rule), the
applier exposes audit-silent *_core methods and the facade keeps the
public reapply-aware wrappers, so the cross-collaborator orchestration
stays in one place (the facade) and the applier remains pure-write.
Promotes _append_exemplar and _find_merchant_by_canonical_name to
append_exemplar and find_merchant_by_canonical_name since the facade's
_categorize_items_inner reaches into them across the module boundary.

No behavior change; full suite passes.
