+++ src/moneybin/services/categorization/__init__.py
@@ -11,7 +11,6 @@ proposal/approval/deactivation lifecycle and depends on this module's
 """
 
 import logging
-import uuid
 from collections.abc import Sequence
 from dataclasses import dataclass
 from time import perf_counter
@@ -21,18 +20,22 @@ import duckdb
 
 from moneybin.config import get_settings
 from moneybin.database import Database
-from moneybin.errors import UserError
+from moneybin.errors import UserError as UserError
 from moneybin.metrics.registry import (
     CATEGORIZE_APPLY_POST_COMMIT_DURATION_SECONDS,
     CATEGORIZE_APPLY_POST_COMMIT_ROWS_AFFECTED,
     CATEGORIZE_DURATION_SECONDS,
     CATEGORIZE_ERRORS_TOTAL,
     CATEGORIZE_ITEMS_TOTAL,
-    CATEGORIZE_WRITE_SKIPPED_PRECEDENCE_TOTAL,
-    MERCHANT_EXEMPLAR_COUNT,
 )
 from moneybin.metrics.registry import (
     CATEGORIZE_MATCH_OUTCOME_TOTAL as CATEGORIZE_MATCH_OUTCOME_TOTAL,
+)
+from moneybin.metrics.registry import (
+    CATEGORIZE_WRITE_SKIPPED_PRECEDENCE_TOTAL as CATEGORIZE_WRITE_SKIPPED_PRECEDENCE_TOTAL,
+)
+from moneybin.metrics.registry import (
+    MERCHANT_EXEMPLAR_COUNT as MERCHANT_EXEMPLAR_COUNT,
 )
 from moneybin.protocol.envelope import ResponseEnvelope, build_envelope
 from moneybin.services._text import (
@@ -41,14 +44,15 @@ from moneybin.services._text import (
 )
 from moneybin.services.audit_service import AuditService
 from moneybin.services.categorization._shared import (
-    SOURCE_PRIORITY,
+    SOURCE_PRIORITY as SOURCE_PRIORITY,
+)
+from moneybin.services.categorization._shared import (
     CategorizationItem,
     CategorizationRuleInput,
     CategorizedBy,
     InternalMatchType,
     Merchant,
     did_you_mean,
-    priority_case_sql,
 )
 from moneybin.services.categorization._shared import (
     MatchType as MatchType,
@@ -58,6 +62,9 @@ from moneybin.services.categorization._shared import (
 )
 from moneybin.services.categorization._shared import (
     matches_pattern as matches_pattern,
+)
+from moneybin.services.categorization._shared import (
+    priority_case_sql as priority_case_sql,
 )
 from moneybin.services.categorization._shared import (
     score_match_shape as score_match_shape,
@@ -71,6 +78,11 @@ from moneybin.services.categorization._shared import (
 from moneybin.services.categorization._shared import (
     validate_rule_items as validate_rule_items,
 )
+from moneybin.services.categorization.applier import (
+    MatchApplier,
+    RuleCreationResult,
+    WriteOutcome,
+)
 from moneybin.services.categorization.matcher import (
     CategorizationMatcher,
     match_merchants,
@@ -78,29 +90,21 @@ from moneybin.services.categorization.matcher import (
 from moneybin.tables import (
     CATEGORIES,
     CATEGORIZATION_RULES,
-    CATEGORY_OVERRIDES,
     FCT_TRANSACTIONS,
     MERCHANTS,
     TRANSACTION_CATEGORIES,
-    USER_CATEGORIES,
-    USER_MERCHANTS,
+)
+from moneybin.tables import (
+    CATEGORY_OVERRIDES as CATEGORY_OVERRIDES,
+)
+from moneybin.tables import (
+    USER_CATEGORIES as USER_CATEGORIES,
+)
+from moneybin.tables import (
+    USER_MERCHANTS as USER_MERCHANTS,
 )
 
 logger = logging.getLogger(__name__)
-
-
-@dataclass(slots=True, frozen=True)
-class WriteOutcome:
-    """Result of a guarded write to ``app.transaction_categories``.
-
-    Returned by :meth:`CategorizationService.write_categorization`. The
-    ``written`` flag distinguishes successful writes (insert or precedence-
-    permitted update) from precedence-blocked attempts. ``skipped_reason`` is
-    populated only when ``written`` is False.
-    """
-
-    written: bool
-    skipped_reason: Literal["lower_priority_source"] | None = None
 
 
 @dataclass(slots=True)
@@ -202,41 +206,6 @@ class CategorizationResult:
         self.errors += len(parse_errors)
 
 
-@dataclass(slots=True)
-class RuleCreationResult:
-    """Typed result for CategorizationService.create_rules."""
-
-    created: int
-    existing: int
-    skipped: int
-    error_details: list[dict[str, str]]
-    rule_ids: list[str]
-
-    def to_envelope(self, input_count: int) -> ResponseEnvelope:
-        """Build a ResponseEnvelope from this rule-creation result."""
-        return build_envelope(
-            data={
-                "created": self.created,
-                "existing": self.existing,
-                "skipped": self.skipped,
-                "rule_ids": self.rule_ids,
-                "error_details": self.error_details,
-            },
-            sensitivity="low",
-            total_count=input_count,
-            actions=[
-                "Use transactions_categorize_rules_list to review all rules",
-            ],
-        )
-
-    def merge_parse_errors(self, parse_errors: list[dict[str, str]]) -> None:
-        """Prepend boundary-validation errors and reflect them in the skipped count."""
-        if not parse_errors:
-            return
-        self.error_details = parse_errors + self.error_details
-        self.skipped += len(parse_errors)
-
-
 class CategorizationService:
     """Canonical categorization surface — merchants, rules, taxonomy, auto-rules.
 
@@ -256,10 +225,9 @@ class CategorizationService:
         self._db = db
         self._audit = audit if audit is not None else AuditService(db)
         self._matcher = CategorizationMatcher(db)
+        self._applier = MatchApplier(db, audit=self._audit)
 
-    # -- Per-transaction category writes (Req 25–31 audit emission) --
-
-    _CATEGORY_AUDIT_TARGET = ("app", "transaction_categories")
+    # -- Per-transaction category writes (audit emission via applier) --
 
     def set_category(
         self,
@@ -270,25 +238,14 @@ class CategorizationService:
         categorized_by: Literal["user"] = "user",
         actor: str,
     ) -> None:
-        """Upsert a transaction's user category and emit ``category.set`` audit.
-
-        Captures the prior row (or NULL) as ``before`` and the new shape as
-        ``after`` so the audit trail can reconstruct overwrites. Mutation +
-        audit row commit atomically.
-        """
-        self._db.begin()
-        try:
-            self.set_category_in_active_txn(
-                transaction_id,
-                category=category,
-                subcategory=subcategory,
-                categorized_by=categorized_by,
-                actor=actor,
-            )
-            self._db.commit()
-        except Exception:
-            self._db.rollback()
-            raise
+        """Upsert a transaction's user category and emit ``category.set`` audit."""
+        self._applier.set_category(
+            transaction_id,
+            category=category,
+            subcategory=subcategory,
+            categorized_by=categorized_by,
+            actor=actor,
+        )
 
     def set_category_in_active_txn(
         self,
@@ -299,90 +256,18 @@ class CategorizationService:
         categorized_by: Literal["user"] = "user",
         actor: str,
     ) -> None:
-        """``set_category`` body without txn boundaries.
-
-        Use when the caller already owns a transaction and wants to batch
-        multiple category writes atomically with their own audit chain.
-
-        ``categorized_by`` is restricted to ``"user"``. The user-manual-edit
-        path is the only intended use; the direct ``INSERT ... ON CONFLICT``
-        below bypasses :meth:`write_categorization`'s precedence guard, so
-        admitting a lower-priority source here would let it silently overwrite
-        a higher-priority existing categorization. New callers needing a
-        non-user write must route through :meth:`write_categorization`.
-        """
-        prior = self._fetch_category_row(transaction_id)
-        self._db.conn.execute(
-            f"""
-            INSERT INTO {TRANSACTION_CATEGORIES.full_name}
-              (transaction_id, category, subcategory,
-               categorized_at, categorized_by)
-            VALUES (?, ?, ?, NOW(), ?)
-            ON CONFLICT (transaction_id) DO UPDATE SET
-                category = EXCLUDED.category,
-                subcategory = EXCLUDED.subcategory,
-                categorized_at = NOW(),
-                categorized_by = EXCLUDED.categorized_by
-            """,  # noqa: S608  # TRANSACTION_CATEGORIES is a TableRef constant
-            [transaction_id, category, subcategory, categorized_by],
-        )
-        after = {
-            "category": category,
-            "subcategory": subcategory,
-            "categorized_by": categorized_by,
-        }
-        self._audit.record_audit_event(
-            action="category.set",
-            target=(*self._CATEGORY_AUDIT_TARGET, transaction_id),
-            before=prior,
-            after=after,
+        """``set_category`` body without txn boundaries (caller owns the transaction)."""
+        self._applier.set_category_in_active_txn(
+            transaction_id,
+            category=category,
+            subcategory=subcategory,
+            categorized_by=categorized_by,
             actor=actor,
         )
 
     def clear_category(self, transaction_id: str, *, actor: str) -> None:
-        """Delete a transaction's category row and emit ``category.clear`` audit.
-
-        No-op (and no audit event) when no row exists.
-        """
-        self._db.begin()
-        try:
-            prior = self._fetch_category_row(transaction_id)
-            if prior is None:
-                self._db.commit()
-                return
-            self._db.conn.execute(
-                f"DELETE FROM {TRANSACTION_CATEGORIES.full_name} WHERE transaction_id = ?",  # noqa: S608  # TableRef constant
-                [transaction_id],
-            )
-            self._audit.record_audit_event(
-                action="category.clear",
-                target=(*self._CATEGORY_AUDIT_TARGET, transaction_id),
-                before=prior,
-                after=None,
-                actor=actor,
-            )
-            self._db.commit()
-        except Exception:
-            self._db.rollback()
-            raise
-
-    def _fetch_category_row(self, transaction_id: str) -> dict[str, Any] | None:
-        """Return the current category row for ``transaction_id`` as a JSON-safe dict."""
-        row = self._db.conn.execute(
-            f"""
-            SELECT category, subcategory, categorized_by
-              FROM {TRANSACTION_CATEGORIES.full_name}
-             WHERE transaction_id = ?
-            """,  # noqa: S608  # TableRef constant
-            [transaction_id],
-        ).fetchone()
-        if row is None:
-            return None
-        return {
-            "category": row[0],
-            "subcategory": row[1],
-            "categorized_by": row[2],
-        }
+        """Delete a transaction's category row and emit ``category.clear`` audit."""
+        self._applier.clear_category(transaction_id, actor=actor)
 
     # -- Merchant lookup / management --
 
@@ -404,136 +289,26 @@ class CategorizationService:
         exemplars: list[str] | None = None,
         reapply: bool = False,
     ) -> str:
-        """Create a merchant mapping.
+        """Create a merchant mapping; optionally fan out to uncategorized rows.
 
-        Args:
-            raw_pattern: Pattern to match in transaction descriptions; pass
-                ``None`` for exemplar-only merchants (``match_type='oneOf'``).
-            canonical_name: Clean merchant name for display.
-            match_type: How to match: 'exact', 'contains', 'regex', or 'oneOf'.
-                Defaults to ``'oneOf'`` — system-created merchants use the
-                exemplar accumulator (categorization-matching-mechanics.md
-                §Schema changes); user-authored merchants pick 'contains' or
-                'regex' explicitly.
-            category: Optional default category for this merchant.
-            subcategory: Optional default subcategory.
-            created_by: Who created the mapping ('user', 'ai', 'rule').
-            exemplars: Initial exemplar set (exact match_text values) for
-                oneOf merchants. Defaults to ``[]``.
-            reapply: When ``True``, runs ``categorize_pending`` after the
-                insert so the new merchant fans out to uncategorized rows.
-                Default ``False`` — callers inside a batch flow (e.g.,
-                ``categorize_items``) skip this and let the enclosing snowball
-                pass do the work instead.
-
-        Returns:
-            The merchant_id of the created merchant.
+        Pure write delegates to ``MatchApplier.create_merchant_core``; when
+        ``reapply=True``, ``categorize_pending`` runs after the insert so the
+        new merchant fans out to uncategorized rows immediately. Callers
+        inside a batch flow (e.g., ``categorize_items``) skip this and let the
+        enclosing snowball pass do the work instead.
         """
-        merchant_id = uuid.uuid4().hex[:12]
-        # DuckDB binds Python lists to VARCHAR[]. An empty list keeps the
-        # column default semantics intact for non-exemplar merchants.
-        exemplars_param: list[str] = list(exemplars) if exemplars else []
-        self._db.execute(
-            f"""
-            INSERT INTO {USER_MERCHANTS.full_name}
-            (merchant_id, raw_pattern, match_type, canonical_name,
-             category, subcategory, created_by, exemplars, updated_at)
-            VALUES (?, ?, ?, ?, ?, ?, ?, ?, CURRENT_TIMESTAMP)
-            """,  # noqa: S608  # USER_MERCHANTS is a TableRef constant, not user input
-            [
-                merchant_id,
-                raw_pattern,
-                match_type,
-                canonical_name,
-                category,
-                subcategory,
-                created_by,
-                exemplars_param,
-            ],
+        merchant_id = self._applier.create_merchant_core(
+            raw_pattern,
+            canonical_name,
+            match_type=match_type,
+            category=category,
+            subcategory=subcategory,
+            created_by=created_by,
+            exemplars=exemplars,
         )
-        if exemplars_param:
-            MERCHANT_EXEMPLAR_COUNT.labels(merchant_id=merchant_id).set(
-                len(exemplars_param)
-            )
-        logger.info(f"Created user merchant {merchant_id}")
         if reapply:
             self.categorize_pending()
         return merchant_id
-
-    def _find_merchant_by_canonical_name(
-        self,
-        canonical_name: str,
-        *,
-        category: str,
-        subcategory: str | None,
-    ) -> str | None:
-        """Return ``merchant_id`` for the oneOf merchant with this canonical name + category, or None.
-
-        Used by the exemplar accumulator to detect whether a previously-created
-        merchant should grow its exemplar set instead of spawning a duplicate.
-
-        Filters on ``match_type='oneOf'`` because exemplars are only consulted
-        for oneOf merchants (see :func:`_match_exemplar`); appending an
-        exemplar to a contains/exact/regex merchant is dead data. Also filters
-        on ``category``/``subcategory``: two ``oneOf`` merchants can legitimately
-        share a canonical name (e.g. an Amazon merchant mapped to ``Shopping``
-        and another mapped to ``Groceries``); merging exemplars across them
-        would cross-pollinate categorization. When the canonical name collides
-        with a different-category merchant, the caller creates a new oneOf
-        merchant alongside it — the two coexist, each doing its own job.
-        """
-        try:
-            if subcategory is None:
-                row = self._db.execute(
-                    f"SELECT merchant_id FROM {USER_MERCHANTS.full_name} "
-                    "WHERE canonical_name = ? AND match_type = 'oneOf' "
-                    "AND category = ? AND subcategory IS NULL "
-                    "LIMIT 1",  # noqa: S608  # TableRef constant
-                    [canonical_name, category],
-                ).fetchone()
-            else:
-                row = self._db.execute(
-                    f"SELECT merchant_id FROM {USER_MERCHANTS.full_name} "
-                    "WHERE canonical_name = ? AND match_type = 'oneOf' "
-                    "AND category = ? AND subcategory = ? "
-                    "LIMIT 1",  # noqa: S608  # TableRef constant
-                    [canonical_name, category, subcategory],
-                ).fetchone()
-        except duckdb.CatalogException:
-            return None
-        return row[0] if row else None
-
-    def _append_exemplar(self, merchant_id: str, match_text: str) -> int:
-        """Append ``match_text`` to a merchant's exemplar set; return new size.
-
-        Idempotent via ``list_distinct``: re-appending an existing exemplar
-        leaves the set unchanged. Updates the per-merchant gauge to surface
-        any merchant whose set is approaching the soft-cap signal (200).
-        """
-        # Single round-trip: DuckDB supports RETURNING on UPDATE, so the
-        # post-update size flows back without a separate SELECT.
-        #
-        # The `updated_at` advance is gated on the exemplars set actually
-        # growing — when the exemplar already exists, the SET expression is
-        # a no-op and the row's per-row freshness must NOT advance, per the
-        # spec's "advances iff a real input changed" contract. See
-        # docs/specs/core-updated-at-convention.md.
-        row = self._db.execute(
-            f"""
-            UPDATE {USER_MERCHANTS.full_name}
-            SET exemplars = list_distinct(list_append(exemplars, ?)),
-                updated_at = CASE
-                    WHEN list_contains(exemplars, ?) THEN updated_at
-                    ELSE CURRENT_TIMESTAMP
-                END
-            WHERE merchant_id = ?
-            RETURNING len(exemplars)
-            """,  # noqa: S608  # USER_MERCHANTS is a TableRef constant
-            [match_text, match_text, merchant_id],
-        ).fetchone()
-        new_size = int(row[0]) if row and row[0] is not None else 0
-        MERCHANT_EXEMPLAR_COUNT.labels(merchant_id=merchant_id).set(new_size)
-        return new_size
 
     # -- Rule management --
 
@@ -545,109 +320,16 @@ class CategorizationService:
     ) -> RuleCreationResult:
         """Create multiple categorization rules in one call (idempotent).
 
-        For each item, looks up an active rule with the same matcher and
-        output (``merchant_pattern``, ``match_type``, ``min_amount``,
-        ``max_amount``, ``account_id``, ``category``, ``subcategory``);
-        ``name`` and ``priority`` are treated as metadata and excluded
-        from the dedup key. If found, returns the existing ``rule_id``
-        and bumps the ``existing`` counter; otherwise INSERTs a fresh
-        12-char UUID hex ``rule_id`` with ``is_active=true`` and
-        ``created_by='ai'``.
-
-        Same matcher with a *different* category output is currently
-        treated as a new rule, not a conflict — see
-        ``docs/specs/mcp-tool-surface.md`` "Rule-conflict detection
-        (follow-up)" for the deferred conflict-resolution work.
-
-        When ``reapply=True``, ``categorize_pending`` runs after the writes so
-        the new rules fan out to uncategorized rows immediately. Source-priority
-        enforcement keeps user manual edits safe regardless.
-
-        Per-row insertion failures are caught so a single bad row does
-        not abort the batch — they appear in ``error_details``.
+        Pure writes delegate to ``MatchApplier.create_rules_core``. When
+        ``reapply=True`` and at least one rule was newly created,
+        ``categorize_pending`` runs so the new rules fan out to uncategorized
+        rows immediately. Source-priority enforcement keeps user manual edits
+        safe regardless.
         """
-        created = 0
-        existing = 0
-        skipped = 0
-        error_details: list[dict[str, str]] = []
-        rule_ids: list[str] = []
-
-        for item in items:
-            try:
-                # DuckDB IS NOT DISTINCT FROM treats NULL = NULL as true,
-                # so optional fields (min/max_amount, account_id, subcategory)
-                # match on NULL across calls.
-                found = self._db.execute(
-                    f"""
-                    SELECT rule_id FROM {CATEGORIZATION_RULES.full_name}
-                    WHERE is_active = true
-                      AND merchant_pattern = ?
-                      AND match_type = ?
-                      AND min_amount IS NOT DISTINCT FROM ?
-                      AND max_amount IS NOT DISTINCT FROM ?
-                      AND account_id IS NOT DISTINCT FROM ?
-                      AND category = ?
-                      AND subcategory IS NOT DISTINCT FROM ?
-                    LIMIT 1
-                    """,  # noqa: S608  # TableRef constant, no user input interpolated
-                    [
-                        item.merchant_pattern,
-                        item.match_type,
-                        item.min_amount,
-                        item.max_amount,
-                        item.account_id,
-                        item.category,
-                        item.subcategory,
-                    ],
-                ).fetchone()
-                if found is not None:
-                    existing += 1
-                    rule_ids.append(found[0])
-                    continue
-                rule_id = uuid.uuid4().hex[:12]
-                self._db.execute(
-                    f"""
-                    INSERT INTO {CATEGORIZATION_RULES.full_name}
-                    (rule_id, name, merchant_pattern, match_type,
-                     min_amount, max_amount, account_id,
-                     category, subcategory, priority, is_active,
-                     created_by, created_at, updated_at)
-                    VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, true,
-                            'ai', CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)
-                    """,  # noqa: S608  # TableRef constant, no user input interpolated
-                    [
-                        rule_id,
-                        item.name,
-                        item.merchant_pattern,
-                        item.match_type,
-                        item.min_amount,
-                        item.max_amount,
-                        item.account_id,
-                        item.category,
-                        item.subcategory,
-                        item.priority,
-                    ],
-                )
-                created += 1
-                rule_ids.append(rule_id)
-            except Exception:  # noqa: BLE001 — DuckDB raises untyped errors on constraint violations
-                skipped += 1
-                logger.exception(f"create_rules failed for rule {item.name!r}")
-                error_details.append({
-                    "name": item.name,
-                    "reason": "Failed to create rule — check logs for details.",
-                })
-
-        if reapply and created > 0:
+        result = self._applier.create_rules_core(items)
+        if reapply and result.created > 0:
             self.categorize_pending()
-
-        return RuleCreationResult(
-            created=created,
-            existing=existing,
-            skipped=skipped,
-            error_details=error_details,
-            rule_ids=rule_ids,
-        )
+        return result
 
     def deactivate_rule(self, rule_id: str, *, reapply: bool = False) -> bool:
         """Soft-delete a rule by setting ``is_active=false``.
@@ -663,25 +345,9 @@ class CategorizationService:
         higher-priority sources (user/migration/ml/plaid) that happen to
         share this rule_id reference are left intact.
         """
-        row = self._db.execute(
-            f"""
-            UPDATE {CATEGORIZATION_RULES.full_name}
-            SET is_active = false, updated_at = CURRENT_TIMESTAMP
-            WHERE rule_id = ?
-            RETURNING rule_id
-            """,  # noqa: S608  # TableRef constant, no user input interpolated
-            [rule_id],
-        ).fetchone()
-        deactivated = row is not None
+        deactivated = self._applier.deactivate_rule_core(rule_id)
         if reapply and deactivated:
-            self._db.execute(
-                f"""
-                DELETE FROM {TRANSACTION_CATEGORIES.full_name}
-                WHERE rule_id = ?
-                  AND categorized_by IN ('rule', 'auto_rule')
-                """,  # noqa: S608  # TableRef constant, no user input interpolated
-                [rule_id],
-            )
+            self._applier.delete_rule_categorizations(rule_id)
             self.categorize_pending()
         return deactivated
 
@@ -694,94 +360,14 @@ class CategorizationService:
         subcategory: str | None = None,
         description: str | None = None,
     ) -> str:
-        """Create a custom user category (active by default).
-
-        Top-level duplicate detection uses an explicit pre-check because
-        DuckDB's UNIQUE constraint treats NULL as distinct. The
-        check-then-insert shape is safe under MoneyBin's single-process,
-        single-writer connection model — see ``database.py`` for the rationale.
-
-        Raises:
-            UserError(code="CATEGORY_ALREADY_EXISTS"): the
-                ``(category, subcategory)`` pair is already present in
-                ``app.user_categories``.
-        """
-        # DuckDB treats NULL != NULL in UNIQUE constraints, so a top-level
-        # category (subcategory IS NULL) can be inserted multiple times without
-        # raising ConstraintException. Guard explicitly for that case.
-        if subcategory is None:
-            existing = self._db.execute(
-                f"""
-                SELECT 1 FROM {USER_CATEGORIES.full_name}
-                WHERE category = ? AND subcategory IS NULL
-                LIMIT 1
-                """,  # noqa: S608  # TableRef constant, no user input interpolated
-                [category],
-            ).fetchone()
-            if existing:
-                raise UserError(
-                    f"Category already exists: {category}",
-                    code="CATEGORY_ALREADY_EXISTS",
-                )
-
-        category_id = uuid.uuid4().hex[:12]
-        try:
-            self._db.execute(
-                f"""
-                INSERT INTO {USER_CATEGORIES.full_name}
-                (category_id, category, subcategory, description,
-                 is_active, created_at, updated_at)
-                VALUES (?, ?, ?, ?, true, CURRENT_TIMESTAMP, CURRENT_TIMESTAMP)
-                """,  # noqa: S608  # TableRef constant, no user input interpolated
-                [category_id, category, subcategory, description],
-            )
-        except duckdb.ConstraintException:
-            sub = f" / {subcategory}" if subcategory else ""
-            raise UserError(
-                f"Category already exists: {category}{sub}",
-                code="CATEGORY_ALREADY_EXISTS",
-            ) from None
-        return category_id
+        """Create a custom user category (active by default)."""
+        return self._applier.create_category(
+            category, subcategory=subcategory, description=description
+        )
 
     def toggle_category(self, category_id: str, *, is_active: bool) -> None:
-        """Enable or disable a category. Existing categorizations are preserved.
-
-        Default categories (is_default=true) write to ``app.category_overrides``;
-        user-created categories update ``app.user_categories.is_active`` directly.
-
-        Raises:
-            UserError(code="CATEGORY_NOT_FOUND"): no category with this ID
-                exists in either ``app.user_categories`` or the seeded defaults.
-        """
-        cat = self._db.execute(
-            f"SELECT is_default FROM {CATEGORIES.full_name} WHERE category_id = ?",  # noqa: S608  # TableRef constant
-            [category_id],
-        ).fetchone()
-        if not cat:
-            raise UserError(
-                f"Category {category_id} not found",
-                code="CATEGORY_NOT_FOUND",
-            )
-
-        if cat[0]:  # default category — record/upsert the override
-            self._db.execute(
-                f"""
-                INSERT INTO {CATEGORY_OVERRIDES.full_name}
-                    (category_id, is_active, updated_at)
-                VALUES (?, ?, CURRENT_TIMESTAMP)
-                ON CONFLICT (category_id) DO UPDATE
-                    SET is_active = excluded.is_active,
-                        updated_at = excluded.updated_at
-                """,  # noqa: S608  # TableRef constant
-                [category_id, is_active],
-            )
-        else:
-            self._db.execute(
-                f"UPDATE {USER_CATEGORIES.full_name} "  # noqa: S608  # TableRef constant
-                f"SET is_active = ?, updated_at = CURRENT_TIMESTAMP "
-                f"WHERE category_id = ?",
-                [is_active, category_id],
-            )
+        """Enable or disable a category. Existing categorizations are preserved."""
+        self._applier.toggle_category(category_id, is_active=is_active)
 
     # -- Categorization core --
 
@@ -796,80 +382,16 @@ class CategorizationService:
         rule_id: str | None = None,
         confidence: float | None = None,
     ) -> WriteOutcome:
-        """Insert or replace a categorization, respecting source precedence.
-
-        Single guarded write path for ``app.transaction_categories``. Lower
-        numeric priority = higher authority (per
-        ``categorization-matching-mechanics.md`` §Source precedence). A new
-        write succeeds only if its source priority is ≤ the existing row's;
-        otherwise the existing row stands and the
-        ``CATEGORIZE_WRITE_SKIPPED_PRECEDENCE_TOTAL`` metric is incremented.
-
-        Returns:
-            ``WriteOutcome.written=True`` if the write took effect (insert or
-            permitted update); ``False`` with ``skipped_reason='lower_priority_source'``
-            if a higher-priority categorization already exists.
-
-        Raises:
-            ValueError: if ``categorized_by`` is not in :data:`SOURCE_PRIORITY`.
-                The generated SQL CASE has no ELSE branch (see
-                :func:`priority_case_sql`), so an unknown source would silently
-                resolve to NULL priority and never overwrite — fail loudly
-                instead of letting a typo masquerade as a precedence skip.
-        """
-        if categorized_by not in SOURCE_PRIORITY:
-            raise ValueError(
-                f"Unknown categorized_by={categorized_by!r}; "
-                f"must be one of {sorted(SOURCE_PRIORITY)}"
-            )
-        # The SQL CASE expression is generated from SOURCE_PRIORITY so the
-        # ladder lives in exactly one place. See priority_case_sql.
-        existing_table = TRANSACTION_CATEGORIES.full_name
-        excluded_priority = priority_case_sql("EXCLUDED.categorized_by")
-        existing_priority = priority_case_sql(f"{existing_table}.categorized_by")
-        cursor = self._db.execute(
-            f"""
-            INSERT INTO {existing_table}
-                (transaction_id, category, subcategory, categorized_at,
-                 categorized_by, merchant_id, rule_id, confidence)
-            VALUES (?, ?, ?, CURRENT_TIMESTAMP, ?, ?, ?, ?)
-            ON CONFLICT (transaction_id) DO UPDATE SET
-                category = EXCLUDED.category,
-                subcategory = EXCLUDED.subcategory,
-                categorized_at = EXCLUDED.categorized_at,
-                categorized_by = EXCLUDED.categorized_by,
-                merchant_id = EXCLUDED.merchant_id,
-                rule_id = EXCLUDED.rule_id,
-                confidence = EXCLUDED.confidence
-            WHERE {excluded_priority} <= {existing_priority}
-            RETURNING transaction_id
-            """,  # noqa: S608  # TRANSACTION_CATEGORIES is a TableRef constant; CASE built from SOURCE_PRIORITY
-            [
-                transaction_id,
-                category,
-                subcategory,
-                categorized_by,
-                merchant_id,
-                rule_id,
-                confidence,
-            ],
+        """Insert or replace a categorization, respecting source precedence."""
+        return self._applier.write_categorization(
+            transaction_id=transaction_id,
+            category=category,
+            subcategory=subcategory,
+            categorized_by=categorized_by,
+            merchant_id=merchant_id,
+            rule_id=rule_id,
+            confidence=confidence,
         )
-        if cursor.fetchone() is not None:
-            return WriteOutcome(written=True)
-
-        # Row exists with a higher-priority source; record the skip with labels
-        # for both sides of the comparison so dashboards can distinguish "ai
-        # blocked by user" from "ai blocked by rule" etc.
-        existing = self._db.execute(
-            f"SELECT categorized_by FROM {TRANSACTION_CATEGORIES.full_name} "
-            "WHERE transaction_id = ?",  # noqa: S608  # TRANSACTION_CATEGORIES is a TableRef constant
-            [transaction_id],
-        ).fetchone()
-        CATEGORIZE_WRITE_SKIPPED_PRECEDENCE_TOTAL.labels(
-            src_existing=existing[0] if existing else "unknown",
-            src_attempted=categorized_by,
-        ).inc()
-        return WriteOutcome(written=False, skipped_reason="lower_priority_source")
 
     def categorize_items(
         self, items: Sequence[CategorizationItem]
@@ -1132,13 +654,13 @@ class CategorizationService:
                 if merchant_id is None and match_text:
                     try:
                         canonical_name = item.canonical_merchant_name or match_text
-                        existing_id = self._find_merchant_by_canonical_name(
+                        existing_id = self._applier.find_merchant_by_canonical_name(
                             canonical_name,
                             category=category,
                             subcategory=subcategory,
                         )
                         if existing_id is not None:
-                            self._append_exemplar(existing_id, match_text)
+                            self._applier.append_exemplar(existing_id, match_text)
                         else:
                             new_merchant_id = self.create_merchant(
                                 None,
