+++ tests/moneybin/test_services/test_categorization_updated_at.py
@@ -98,7 +98,7 @@ class TestUserMerchantsUpdatedAt:
         # _append_exemplar is an internal write site exercised here directly
         # so the updated_at refresh is pinned at the service boundary, not
         # several layers up through a categorization batch.
-        svc._append_exemplar(merchant_id, "STARBUCKS #5678")  # pyright: ignore[reportPrivateUsage]
+        svc._applier.append_exemplar(merchant_id, "STARBUCKS #5678")  # pyright: ignore[reportPrivateUsage]
         after = db.execute(
             "SELECT updated_at FROM app.user_merchants WHERE merchant_id = ?",
             [merchant_id],
@@ -129,7 +129,7 @@ class TestUserMerchantsUpdatedAt:
         ).fetchone()
         assert before is not None
         time.sleep(0.01)
-        svc._append_exemplar(merchant_id, "STARBUCKS #1234")  # pyright: ignore[reportPrivateUsage]
+        svc._applier.append_exemplar(merchant_id, "STARBUCKS #1234")  # pyright: ignore[reportPrivateUsage]
         after = db.execute(
             "SELECT updated_at FROM app.user_merchants WHERE merchant_id = ?",
             [merchant_id],
